avoid code duplication of fromEncodedComponent_helper(QGIS)/qt_urlRecodeByteArray(Qt)
