+++ python/core/auto_generated/qgis.sip.in
@@ -10,6 +10,7 @@
 
 
 
+
 %ModuleHeaderCode
 #include <qgis.h>
 %End
